mejora en el delete de img

+++ clientes/models.py
@@ -2,6 +2,8 @@ from django.db import models
 import os
 
 # Create your models here.
+
+
 class Cliente(models.Model):
     nombre = models.CharField(max_length=100)
     apellidos = models.CharField(max_length=100)
@@ -12,8 +14,14 @@ class Cliente(models.Model):
         return f"{self.nombre} {self.apellidos}"
 
 # Eliminar imagen cuando se elimina cliente
+# ELIMINAR FOTO AL MOMENTO DE ELIMINAR CAMPO
+
+
 def delete(self, *args, **kwargs):
-    if self.foto:
-        if os.path.isfile(self.foto.path):
+    if self.foto and os.path.isfile(self.foto.path):
+        try:
             os.remove(self.foto.path)
+        except OSError as e:
+            # Manejar errores, por ejemplo, escribir en el log
+            print(f"Error al eliminar el archivo: {e}")
     super().delete(*args, **kwargs)
